update sample set creation and migrator to use `sample set {samp set count}` pattern for naming
